enable CLI autodoc. Refactor arg parse to separate file to avoid import weirdness during doc generation
